move all code to src + remove code folder
