+++ Makefile
@@ -1,2 +1,2 @@
-test-all:
+test:
 	python -m unittest discover
