+++ tests/test_molecule.py
@@ -1,8 +1,9 @@
 import unittest
-from code.helpers.constants import MOLECULES_SMILES
-from code.Molecule import Molecule
 
 from rdkit import Chem
+
+from src.helpers.constants import MOLECULES_SMILES
+from src.Molecule import Molecule
 
 
 class TestMolecule(unittest.TestCase):
